Fix missing index on magnetic fields
The issue was that we were missing the right most face-centered magnetic
field. In general to set that we need the boundary conditions. To fix
this I made the Grid1D.magnetic vector an extra index long and tweaked
the boundary conditions and initial conditions to set the last magnetic
face correctly.